Move tastyBootstrap to a new test file
The test "tastyBootstrap" is different to the rest of the compilation
tests because it actually recompiles the compiler sources, which can
interfer when using CompilationTests.
